+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project attempts to match the major and minor versions of [stactools](https://github.com/stac-utils/stactools) and increments the patch number as needed.
 
+
+## [v0.0.0]
+
+### Added
+
+- Functionality to support the split of the ecmwf-forecast collection into 7 separate collections (ecmwf-forecast-wave-fc, ecmwf-forecast-cda-fc, ecmwf-forecast-oper-fc, ecmwf-forecast-enfo-ep, ecmwf-forecast-waef-ef, ecmwf-forecast-waef-ep, ecmwf-forecast-scwv-fc)
+- Functionality to add kerchunk indices to the STAC item
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
+
 ## [Unreleased]
 
 ### Added